+++ packages/ui-core/CHANGELOG.md
@@ -66,7 +66,7 @@
     - correct header css for style guide, delete h6 props
     - updated documentation
     - synchronized colors according to the guide
-    - add hAlign prop
+    - add hAlign prop width default value 'inherit'
     - fix children type
 - **cn**
     - add variability to add arguments